refactor: add unit tests for RecentSearchDao methods
